consolidate shared api hooks
